Always send expired notice when an account expires

+++ includes/member-functions.php
@@ -486,7 +486,9 @@ function rcp_set_status( $user_id, $new_status ) {
 
 	if( update_user_meta( $user_id, 'rcp_status', $new_status ) ) {
 
-		delete_user_meta( $user_id, '_rcp_expired_email_sent');
+		if( 'expired' != $new_status ) {
+			delete_user_meta( $user_id, '_rcp_expired_email_sent');
+		}
 
 		do_action( 'rcp_set_status', $new_status, $user_id );
